+++ CHANGELOG.md
@@ -1,2 +1,8 @@
-owncloud-onlyoffice (1.0.0)
-* Initial release
+## 1.0.1
+- fix exception when versions app is disabled
+- adding protocol to document server url
+- onlyofficeOpen is default action
+- Nextcloud 11 compatibility
+
+## 1.0.0
+- Initial release